Cassandra mutate cleanups and timestamp changes
The timestamps used by the Cassandra-Thrift KCV store now approximate
the number of nanoseconds since the UNIX Epoch.  I estimate the
imprecision in this approximation as up to about one millisecond of
slowness wrt true time.  The error bound can be tuned downward with
increasingly fancy static initializer code.

The final bit in any timestamp is 0 for delete mutations and 1 for
insert mutations.  This limits effective time precision in timestamps
to 2 ns increments.

The mutate and mutateMany methods in the Cassandra-Thrift KCV store
have been reimplemented to make just one batch_mutate call to the
Cassandra backend.  The obsolete insert, delete, insertMany, and
deleteMany methods are gone.  The only public methods besides the ctor
in this class are now @Overrides of some current interface method.