Refactored and documented locking/idalloc Rid generation
This adds javadoc for diskstorage.util.ConfigHelper#getRid().
It also tweaks how Rid is guessed when the user has provided no
Rid-related configuration keys (the default case).  This goes back
to using InetAddress.getLocalHost and RuntimeMXBean's JVM "name",
which is typically the hostname and the operating system pid of the
JVM.  However, the javadoc makes no guarantees about the actual
value of "name" in general, so this is broken from a method contract
standpoint.