Enabled HBase locking and id allocation tests
The HBase implementation passes the HBase lock and ordered key-col-
value store tests on my machine using HBase 0.92.1.

This commit also makes the HBase OKCVS interpret zero-length getSlice
bounds as "no limit", in the fashion of Cassandra.  I'll probably
change the behavior of getSlice to only interpret a null bound as
"no limit" and add logic in the Cassandra OKCVS to convert those
nulls to a reference to a static zero-length ByteBuffer.