Added Cassandra daemon embedding for testing
The tests formerly known as CassandraThriftKeyColumnValueTest,
CassandraLockKeyColumnValueStoreTest, and CassandraGrapTest are now
prefixed with "External".  This signifies that they start Cassandra in
a process external to the JVM process for testing purposes.  Also
added are the same classes with "Internal" in place of "External";
these use a newly-added thin wrapper around the
org.apache.cassandra.thrift.CassandraDaemon class to start a Cassandra
instance inside the JVM executing the tests.

pom.xml also excludes the newer Avro artifact required by HBase in
favor of the older one required by Cassandra (this conflict caused
fatal exceptions when trying to run an embedded Cassandra instance as
described in the previous paragraph).  This might have broken HBase
support.  We will almost certainly have to revisit this temporary
exclusion later.