HBase: fix mutateMany timestamps
Replaced hard-coded timestamp value of 0L (and the TODO comments
alongside them) with timestamps generated from
System.currentTimeMillis().  Specifically, the timestamp for all
deletions is equal to the value returned by System.currentTimeMillis()
early in mutateMany, and the timestamp for all puts is that plus one.
Just before returning, the method checks wthere
System.currentTimeMillis() returns a value higher than the put
timestamp; if not, it does Thread.sleep(1L) calls until this is so.