Added alpha-quality embedded Cassandra backend
The CassandraEmbedded StorageManager and OKCVS implementations talk to
Cassandra using ordinary method calls instead of Thrift.  Cassandra
must already be running in the JVM for this to work.  The four unit
tests added by this commit exercise the embedded backend.  The only
testing I've done on this code is the unit tests in this commit.

This backend is essentially a stripped-down adaptation of the code
sitting between Thrift and StorageProxy in Cassandra mainline.
Permission checks and keyspace/column family validation checks present
in mainline have been removed in this implementation.