Fix backend on embedded-cassandra graph test
InternalCassandraEmbeddedGraphTest was using the ordinary
Cassandra-Thrift backend instead of the embedded backend.  It wasn't
testing the embedded backend at all.  This commit makes the test use
the intended Cassandra-Embedded backend.

This oversight only appears to have affected the graphdb-level test.
The tests which exercise Cassandra-Embedded backend at the
diskstorage-level are in good shape.