Added autostart to embedded-Cassandra backend
When a newly-added CassandraEmbeddedStorageManager configuration
property is set, the storage manager constructor will start a
Cassandra thread in the background using the cassandra.yaml file
pointed to by the config property.  This is internally synchronized to
avoid starting multiple Cassandra threads in a single JVM, so
subsequent constructor invocations can't start additional Cassandra
threads even when this config property is set.