LocalLockManager tweaks
This commit is a bundle of changes copied over in patch form from
another workstation.  I should probably be using git-bundle.

LocalLockManager is in an intermediate state in this commit: it now
polices zombie locks, but it also permits attempts to lock with
expiration times in the past.  The unit test is also anemic.

I'm planning to add a unit test at the KeyColumnValueStore level which
explicitly configures lock timeout calls acquireLock() on a
transaction and sleeps to test it.