Added @SuppressWarnings("unused") to scanTest()
The comment in KeyColumnValueStoreTest's scanTest() method indicates
that it intentionally creates but does not use the last RecordIterator
that it requests from the storage layer.  KeyColumnValueTest had the
same pattern but no comment and no warning suppression, and I had
"fixed" it by adding assertions on it.  This commit reverts my change
in KeyColumnValueTest and additional protects the assignment in
question with a comment and @SuppressWarnings("unused")