Type fix in Astyanax getSlice()
The Astyanax class RowQuery<K, C> has two colliding/ambiguous method
signatures for withColumnRange() when K=C=ByteBuffer.  I'm using the
raw type to avoid this problem, but it's an unclean solution and
prevents the compiler from typechecking the first two arguments to
withColumnRange.  These used to be ByteBuffers but were changed to
StaticBuffer in a recent commit, causing Astyanax to die.  This patch
converts the StaticBuffers back to ByteBuffers.  I'd still rather find
a way to let the compiler typecheck the method call in question
without running aground on ambguous signatures.