Fix Astyanax slice end boundary handling
The Astyanax becakend had an equals() call comparing StaticBuffer and
ByteBuffer.  Because the concrete types differed, this call
unconditionally returned false.  Fixed to compare objects of the same
concrete type.

This failing equals() check caused the end boundary in Astyanax
getSlice() calls to be erroneously included in the results when it
should have been excluded.