Change WeakHashMap to ConcurrentHashMap
The WeakHashMap in TitanBlueprintsGraph can be accessed concurrently
by multiple threads.  This access pattern is tested in
TransactionalGraphTestSuite#testCompetingThreads().  The javadoc for
WeakHashMap says the collection is unsynchronized.  When running the
test on my box, these unsynchronized concurrent accesses seem to the
type of failure seen in #284.

I changed the WeakHashMap to a ConcurrentHashMap.
Collections.synchronizedMap(new WeakHashMap()) would also work, but I
suspect that we want concurrency more than we want the ability to
automatically forget and GC uncommitted/unrollbacked transactions.