Relax lock timestamp mismatch check
Writing a lock claim to storage may appear to fail on the client,
generating a StorageException, even though it actually succeeded.  For
example, a Cassandra cluster could complete a write but fail to
acknowledge it before triggering a TimeoutException.  These claims
linger in storage and pointlessly block lock aquisition by other
processes until aging past the lock expiration time.

This commit changes the lock verification step to accept a sequence of
claims sharing the same rid and differing only in their timestamps, so
long as the final timestamp matches that expected by the verifier.
This should alleviate one potential cause of timeout mismatch
exceptions as discussed in #302.