Added Cassandra test profiles
Two default test-runner profiles trigger in titan-cassandra's pom.xml
when the -Dtest=... argument is absent on the command line.  A third
profile triggers when -Dtest=... is present.

The sole objective of this commit is to make maven stop running a
single test specified with the -Dtest=... command line option twice.
There might be a more elegant way to do it, but this works on my box.