Hacked an Astyanax-Recipes-based lock impl
This is a proof-of-concept, but it would need cleanup in titan-core's
configuration of locking Backend.java, in the configuration of graphdb
tests using Astyanax, and also would benefit from a common abstract
base class providing functionality duplicated across
AstyanaxRecipeLocker and ConsistentKeyLocker.  I moved towards the
last point making LockState parameterized in the type of its
LockStatus class, but there is still work to be done on that point and
more on the other two.

The following three tests passed on my box with the new
AstyanaxRecipeLocker in use (but the class is otherwise totally
untested and should be presumed buggy):

* InternalAstyanaxGraphTest.java
* InternalAstyanaxGraphConcurrentTest.java
* InternalAstyanaxLockKeyColumnValueStoreTest.java