Made script to update gh-pages at release time
When the titan-site maven module is packaged with -Ptitan-release, it
will write a script target/site-scripts/gh-pages-update.sh.  This
script takes as its sole argument a release.properties file generated
by mvn release:prepare.  The script checks out a new copy of the repo
on branch gh-pages, then locally stages new javadoc, wikidoc, and a
new root index.html appropriate for the release version (the "release
version" is read from release.properties).  The script adds all its
changes to the git index but does not commit or push, leaving the
chance for the user to inspect and potentially modify its output.