Tweaks for timeout-sensitive failure on Travis
Travis builds 227.1 and .3 failed due to the same exception in the
same test method, parallelNoncontendedLockStressTest.  The log
suggests that locks are expiring before the LockStressor threads in
this test can check their locks, resulting in an underrun on a success
counter and the assertion shown in the logs.

I made two changes:

* Changed expire exception type from PermanentLockingException to
  TemporaryLockingException and made the test count the sum of
  successful locked operations and operations that failed due to
  TemporaryLockingExceptions (but not Permanent or any other type)

* Increased the test's lock expire time from 1s to 5s

I also added a log statement that might help check my analysis if the
build fails again on Travis.  This commit passed the test suite
locally, but so did the commit that just failed on Travis.