Removing IDAuthority's peekNextID method
In the consistentkey backend, this peek method could skip valid
blocks, limiting its usefulness.  It deosn't appear to be used
anywhere outside a single test method, so I deleted the method from
both its interface and implementations.

In the consistentkey implementation of IDAuthority, the process of
allocating a block happens in approximately two steps.  1. Write the
backend to claim an ID block.  2. If the write took longer than a
certain configurable time limit, try to delete it, otherwise return
the block as successfully allocated.

The problem: "peek" just looks at written claims.  If peek happens
after a write completed but before the responsible machine deletes it,
then peek will consider it allocated and skip it.  A subsequent call
for the next block may then return a start less than what peek
reported.

If we come up with a genuine need to peek at ID blocks, then we'll
need to modify the consistentkey implementation to avoid this behavior
(unless we can live with it, somehow).