Rewriting Titan distro scripts
Replaced titan.sh with a script that starts or stops Cassandra
followed by Rexster with a Titan+Cassandra graph configuration.  Added
Cassandra and Rexster config files and logging config files that write
within the distro's extracted directory.  The new titan.sh can also
delete these data when invoked with the "clear" argument.

This commit is a work-in-progress and should not be packaged or
shipped.  This commit leaves the HBase and BDB distros in a broken
state.  A simplified titan.sh and set of rexster.xml config files is
needed for HBase and BDB.