Customizing imports in zip/tarball Rexster cfg
This shows that we can mirror the customized imports in Titan's
Console.java file in Rexster without modifying Rexster's source or
resorting to an extension.  However, this commit introduces
copy-pasted duplication between Console.java and rexster.xml.  I think
I see a better way to do this now.  The logic in Console.java that
executes groovy import statements should be extracted into a static
method or a method on some new throwaway object.  It will return
groovy code.  A new main method should be introduced to call this
helper method and write the code to a file.  Rexster can then be
configured to execute it via the <init-scripts> configuration option
in the <script-engine> stanza for Gremlin.  It might even be possible
to build this into the titan-dist maven lifecycle, so that it's fully
automated.