Moved performance test tuning into JUBProvider
Logic to estimate JUB test method benchmark rounds required to run for
a total of 5 seconds, save those multipliers to a file, read them, and
modify the JUnitBenchmarks configuration accordingly is now
concentrated solely in JUnitBenchmarkProvider, instead of being split
across JUnitBenchmarkProvider and GroovyTestSupport.  The
implementation uses java.lang.reflect.Proxy to generate a dynamic
proxy annotation instance and intercept JUB's call to the
benchmarkRounds() method for each test.  This is sort of byzantine,
but it's the best I can do, given the severe limits on inheritance in
both JUnit and JUnitBenchmarks via the final keyword and restrictive
method visibility modifiers.