Extracted Gremlin imports into utility class
Console.java now relies on a new class, ConsoleSetup.java, to supply a
list of imports/import-statics to execute when bootstrapping a Gremlin
shell for Titan.  The utility class also contains a main method that
writes the imports to stdout or a file.  This is useful for generating
a Rexster <init-script> file for use with Titan that imports all the
same symbols as an interactive gremlin.sh session.  It also avoids
copy-pasting, so that any time we update the import list effective in
gremlin.sh, we will also automatically update the Rexster environment.