Remove bin/*.local files from distro archives
This commit deletes some Titan configuration files from the bin
directory of zip/tar archives built by titan-dist and its submodules.
In lieu of these unusually named files in the bin directory, an
expanded set of config files copied from the /config directory of the
repo is now copied to the /conf directory in the zip/tar achives.