Moving titan-dist-* submodules into their parent
I'm trying to reduce the number of archive modules in Titan.  We had
almost as many as we did actual source code modules, spreading out our
distro-related files over a painfully disordered directory structure.
Trying to centralize everything in a single module using profiles in a
way that won't break the release plugin.  Not well tested.