Renaming Cassandra compression config keys
Removed the "cassandra.db." prefix from these Cassandra configuration
keys:

* compression.enabled
* compression.chunk_length_kb
* compression.sstable_compression

Also: the middle option in the list was formerly chunk_size_kb.  I
changed it to chunk_length_kb because that's the string Cassandra uses
and I don't see a reason why the need to be subtly different.

I think the "cassandra" prefix is redundant because graph configs have
"storage.backend", which must be set to a one of the strings
identifying a Cassandra backend for any of these keys to be read by
Titan in the first place.  "db" is debatable.