Make BytesToken format hack safe on other tokens
Check that objects of general type Token<?> pass an instanceof
BytesToken check before trying to apply a hack for
https://issues.apache.org/jira/browse/CASSANDRA-5566

I think it was impossible for the hack to have fired prior to this
commit.  It was guarded by a regex on Token strings that looked for
"Token(bytes[.+])".  Murmur3- and RandomPartitioner return LongToken
and BigIntegerToken, respectively, whose toString() methods both
return strings containing only decimal digits and a minus sign.

Still, I think the check is useful in case somebody has a third-party
partitioner that uses tokens returning arbitrary strings.