Hacking HazelcastKCVS to check column slices
Why did getSlice explicitly use a signed bytes comparator when testing
columns against the lower and upper slice bounds?  It's clear that
this was deliberate, but I can't see the reasoning behind it.  It
makes a current getSlice test fail.

I changed that to unsigned and also implemented a naive column bounds
check in getKeys.  This getKeys hack could be smarter about the
caching columns it reads during slice checking for potential later
reuse by getEntries, but I'm sticking with the simple-and-dumb version
initially.