A new Expect-based release archive test
I'm trying this out.  When running "mvn verify -Paurelius-release" in
titan-dist-berkeleyje, an integration test extracts the distribution
zipfile and runs an Expect script that spawns gremlin.sh and runs a
few trivial commands (adds a vertex and commits it).  If any of the
expect subcommands timeout or don't see what they want, then the
script exits with nonzero status and the integration test fails, in
turn failing the verify lifecycle stage for the module.