Make HC transactions use LOCK_EXPIRE_MS
Hazelcast's notion of expiring transactions is not really identical to
Titan's notion of expiring locks.  Might want to introduce a new
config key instead of reusing LOCK_EXPIRE_MS.

This commit might fix a cause of #425.  This issue seems to be timing
sensitive and I've yet to reproduce it, so I can't tell for sure.