Copy prerelease Furnace generators into titan-test
titan-test uses com.tinkerpop.furnace.generators, but Furnace has yet
to be released.  Only SNAPSHOT versions are available.  We can't
release Titan with a SNAPSHOT dependency on Furnace.  As a hack, I
copied the source for the generators subpackage into titan-test's
source tree and changed the package name to
com.tinkerpop.furnace.alpha.generators.  The generators subpackage
happens to be self-contained and depends on no other Furnace files.

I commented the Furnace SNAPSHOT dependency in titan-test/pom.xml.

This should all get reverted when Furnace releases and we can depend
on it in releases.