Tweaks to satisfy the release plugin
The release:prepare goal solicits a new version for titan-dist-parent,
but because it only appears through pom inheritance and not through a
module relationship, it appears to do nothing with that new version,
leading to an error in release:prepare's invocation of the verify
phase.  This hack attempts to trick release:prepare into adding
titan-dist-parent as a submodule of the toplevel pom.