Delete gh-pages profile from titan-site
I used this profile to experiment with the Github maven plugin, but it
proved too destructive in testing to use in actual releases.  It
doesn't appear to offer an easy way to replace a subdirectory with a
new one a la rsync, so its tricky bordering on impractical to use it
to modify subdirectories of gh-pages's javadoc and wikidoc
directories.  All its prospective titan-site functionality has been
subsumed in the gh-pages-update.sh script anyway.