Activate mem/perf test profiles using properties
Profile activation with -P... has proven brittle and limited because
of the way it interacts with <activeByDefault> (disables all marked
profiles in the same pom but not parent or child poms).  I'm changing
the memory-test and profile-test profile activation criteria.  Now
they activate when -Dtest.mem=true or -Dtest.perf=true are present on
the command line (respectively).  Using -Pperformance-test or
-Pmemory-test is not recommended.