Cleanup titan-cassandra/pom.xml
This commit removes unreferenced properties and reduces the overall
size of the pom by moving the BOP/RP Cassandra test configurations out
of profiles and into the <build> tag of the pom.  Both the BP and RP
tests are enabled by default.  To skip one, set
-Dtest.skip.ordered=true or -Dtest.skip.unordered=true (respectively).
I would rather these properties be named test.ordered and
test.unordered and their boolean sense be negated, but I don't see an
easy way to do it with Maven right now.