Make titan.sh wait for Cassandra's Thrift server
Added nodetool to titan-server/bin/.  Instead of starting Cassandra
and sleeping for a hardcoded interval, titan.sh now starts Cassandra
and then calls `nodetool statusthrift` and waits up to 60 seconds
until that command returns the string 'running'.  If it fails to
return that string within the timeout, then titan.sh prints an error
to stderr and exits without attempting to start Rexster.