Fix Rexster log4j configuration
Rexster's Application.java loads "log4j.properties" in a static
initializer and does not seem to provide a way to override this
behavior through configuration.  So, I've created the directory
"rexhome" within the root of the titan-server archive containing
log4j.properties for rexster (and a "public" directory for rexster
containing web assets).  Now rexster logs to
$ARCHIVE_ROOT/log/rexstitan.log.