Infinispan: added SFS persistence config options
This commit adds support for Infinispan's SingleFileStoreCache
persistence mechanism.  I've only tested it in a local,
non-distributed configuration with sample data, but it seems to work
as intended there.  I have not scrutinized the async and preload
options to see whether they actually function as documented, but the
configuration code involved is pretty trivial compared to the
transactional stuff.

This commit also wires the StaticBufferAE AdvancedExternalizer
implementation into Infinispan's GlobalConfiguration by default.  I
tested this a few days back on EC2 when doing replicated cluster
testing, and I'm reasonably certain it does the right thing under
favorable network conditions and the absence of node failure, but it's
not tested beyond that point.  I had to wire in the externalizer to
get persistence running.