Fixing timestamp handling bug in CK locker
ExpectedValueCheckingStore and ConsistentKeyLocker still expected
their store.mutate(...) calls to use the current timestamp.  Since the
timestamp is now controlled by StoreTxConfig, the locking code wound
up writing all its mutations within a tx with the same timestamp.
This led to thousands of warnings when testing EVCS and CKL on top of
titan-cassandra.

This commit makes each mutate call use a fresh timestamp.

titan-cassandra also recently got a new performance-regression test in
the form of TitanEventualGraphTest#testCachePerformance().  This isn't
part of @Category({ PerformanceTests.class }).  It's intended to be
run as part of the regular test suite as a "sanity check" on the
performance of the cache subsystem.  But titan-cassandra executes its
tests in multiple threads, so while the test passed in isolation, it
would sometimes randomly fail when other threads won the scheduler
wars.  Now this test has its own @Category({ SerialTests.class }) and
it runs in its own single-threaded surefire execution.