Make HBase shutdown more carefully in tests
When running memory and performance tests through Maven, I found that
HBase would die after perf tests but fail to restart for mem tests.

This turned out to be ZK failing to bind its port.  The HBase test
harness killed the HBase process with SIGTERM, which seems to prevent
ZK from cleanly closing its listen socket, which then goes into
TIME_WAIT and prevents the next ZK process from binding the port.

To fix this, I made the test harness first try to stop HBase using the
clean shutdown script provided with the HBase distro, waiting up to 20
seconds for it to finish.  If it doesn't finish, the harness resorts
to SIGKILL (not TERM).

This approach seems to resolve the issue -- now perf and mem tests
both run successfully.