Reducing iteration count in testTranactionalMemory
A single execution of this method takes approximately 35 minutes on an
EC2 m1.medium.  This is unnecessarily for performance testing purposes
with JUnitBenchmarks.

I reduced the iteration count by an order of magnitude (100k to 10k)
and made the thread count equal to the number of cores * 2 rather than
the hardcoded number 8.