Expunge uses of `uname -o` from shell scripts
The -o option is not POSIX standard and does not exist on OS X (and
possibly other platforms).  Replaced with calls to just `uname`,
without options, which under POSIX must behave the same as `uname -s`.

This change is untested.  It's intended to address #534, but needs to
be verified manually before closing the issue.