Wrap QueryExecutor implementations with Metrics
This times and counts (invocations and exceptions counted separately)
calls to the methods on implementations of the QueryExecutor
interface, when the calling transaction has metrics enabled.  Any
transaction-specific metrics prefix, if configured, is honored.  Time
spent in iterators returned by methods on the QueryExecutor interface
are *not* covered by these metrics.  This is part of #402.