Update ConsistentKeyLockerTest for Metrics changes
ConsistentKeyLockerTest is, by design, extremely sensitive to
internals changes in ConsistentKeyLocker or AbstractLocker.  The test
started failing when I added calls to
tx.getConfiguration().getMetricsPrefix() in AbstractLocker.  In this
commit, I changed ConsistentKeyLocker to allow any number of calls to
tx.getConfiguration().  The configuration doesn't hold mutable lock
state, and this is the second time I've stumbled over spurious
tx.getConfiguration() failures in ConsistentKeyLockerTest, so I think
relaxing this constraint is appropriate.