Make debian package rules use mvn --also-make
Changing into the titan-dist/titan-dist-all subdirectory and
attempting to build only that module fails if its peer and parent
dependencies aren't already installed in the local repository.  The
debian package rules script used this technique and would fail on a
pristine local Maven repo.  This commit uses Maven's -pl/--projects
and -am/--also-make options to build titan-dist/titan-dist-all and all
of the modules on which it dependends (but nothing more).