Refactor, integrate, test expired lock cleaner
The lock cleaner is now integrated with core.  It's controlled through
a new GraphDatabaseConfiguration option, "lock-clean-expired".  It's
disabled by default.  This commit adds/expands tests on the three
classes involved.  These changes seem to work in unit and limited
manual testing on my workstation, but the code is all very new and may
be buggy.