Groundwork for configuring timestamp resolution
Turn TimestampUtility into a general interface with a
nanosecond-resolutiond and microsecond-resolution implementation.
This change removes a few points where TimeUnit.NANOSECONDS or a
something-to-nanoseconds conversion divisor was hardcoded.  The
titan-test and titan-cassandra suites pass.  The other test suites may
be broken.  Also, I'm not sure these changes are complete -- I have
yet to turn on microsecond timestamps and see what breaks.

This relates to #472 and #483 but does not close either.