Add general sleep-after-mutate method
HBase already slept after mutating, but now instead of sleeping for a
hardcoded millisecond it sleeps for a time based on its
TimestampProvider (but falls back to 1 ms if no provider is set).

Cassandra could get away without this method in its mutate
implementations back when it used nanotime, but with millisecond
resolution (Timestamps.MILLI or .MICRO providers), some of the tests
routinely fail because multiple operations are colliding inside a
single millisecond (MultiWrite especially).