Make IDPoolTest parallel
IDPoolTest creates an array of Threads.  It then iterates over the
array, first calling run (not start), and the iterates a second time
calling join.

Because the context contains joins on all these threads, I suspect
calling run was a mistake and start was intended.  I've modified the
line in question.