Refactor ID Partitioning configuration
Instead of a boolean, ID partitioning is now a three-valued enum:

* DEFAULT
* ENABLED
* DISABLED

ENABLED corresponds to the old value of true, and DISABLED corresponds
to the old value of false.  DEFAULT is the new behavior.  This enables
partitioning if the backend store reports isKeyOrdered() &&
isDistributed() via its StoreFeatures.  Otherwise partitioning is
disabled.

A warning is logged if the user sets DISABLED but the DEFAULT feature
criteria are true.