Handle zero upper partition bounds
VertexIDAssigner reads the first four bytes of key ranges provided by
getLocalKeyPartition as ints.  It then unsigned right shifts the ints
by two bits each to make them effectively 30 bits wide.  It does some
arithmetic on the ints to preserve inclusive-start and exclusive-end
semantics and then passes them to PartitionIDRange's constructor.

Here's the computation done on each range-end int:

  partition[1] = (partition[1] >>> 2) - 1;

If the original partition[1] value is 0, this results in
Integer.MIN_VALUE.  Because the partition width is 30 bits, this value
is invalid (too wide since the MSB is set).  This value leads to a
precondition failure shortly after in PartitionIDRange's constructor.

I think the intent is to follow the subtraction by zeroing the two
most significant bits, simulating 30-bit wide unsigned integer
arithmetic and allowing underflow.  So this commit adds a bitmask
after subtraction like so:

  partition[1] = (partition[1] >>> 2) - 1;
  partition[1] &= 0x3FFFFFFF;