Refactor HBase connection handling
Now HBase manager is 1-to-1 with HConnection.  Each manager creates
its single HConnection during construction and assigns it to a final
field.  However, every method is responsible for calling
connection.getTable(String) and closing the returned table object in a
finally block.  These HTableInterface instances on the stack are not
thread safe.  The HConnection is only closed when the manager is
closed.

Compared to creating an HConnection on the stack each method, this
approach is faster and does not cause too many open files errors
during testing.