Skip backoff when exhausting ID uniqueid
Uniqueid exhaustion within a partition has nothing to do with backend
workload.  At most, we should wait the default backoff before
retrying; an increasing backoff is definitely not warranted.  This
commit removes any sort of backoff after an ID allocation attempt in
which all storage backend requests succeed but the uniqueid turns out
to be exhausted.