Fix synchronization in LogTest#testFuzzMessages
This commit also hard-codes the length and number of messages in the
log fuzz test, leaving only the byte values to vary pseudorandomly.
The BDB edition of LogTest fails when I set 32 rounds of 512 byte
messages (or larger), but succeeds with 32 x 256.  The other backend
subclasses of LogTest don't have this problem.  More investigation
needed.