Flip broken VertexIDAssignerTest partition bounds
Fixes #591

VertexIDAssignerTest uses a MockIDAuthority that takes its partition
bounds as an integer array of length 2, then writes the ints to
StaticBuffers in its getLocalIDPartition() method.  A precondition in
this method checks that the lower buffer compares before the upper
buffer.

VertexIDAssignerTest is a JUnit parameterized test, and one parameter
set passed the int array { -10000, 10000 } to MockIDAuthority.  This
was guaranteed to fail the precondition mentioned above, since the
negative int has the MSB set and the positive int doesn't.  This
showed up as an IllegalArgumentException in
titan-test/target/test.log.

I think this bug was limited to MockIDAuthority and
VertexIDAssignerTest.  I don't see how it could affect
VertexIDAssigner or ConsistentKeyIDManager.  CKIM works in terms of
opaque bytes.  VertexIDAssigner does work with ints, but it
unsigned-shifts them down two bits before doing any comparisons,
discarding the lower order bits and guaranteeing that the resulting
int values are positive.