Increase timeout in #631
Attempting to punt this per discussion with @mbroecheler; not fixing
the root problem, just increasing the timeout value involved in the
test logic's race condition

A slightly more labor-intensive but more-reliable approach might be to
make the test block up to a much longer timeout (order of magnitude
longer than median runtime), but wakeup frequently to test whether its
expected conditions have been met and bail out early if so.

The right way to do this that's a lot more work is to mock
TimestampProvider and completely eliminate Thread.sleep calls in the
test logic.