Port Cassandra 2.0.6 script/conf bits to Titan zip
Updated the Cassandra configuration and shell script files we include
in the titan-$VERSION.zip file with changes from Cassandra 2.0.6.
I've retained Titan-specific modifications to bin/cassandra and
bin/nodetool that force use of the cassandra.in.sh file included in
the zip rather than any that might be present in /etc or the user's
homedir, but otherwise these files should now reflect their
counterparts in upstream as of 2.0.6.