Initial AsciiDoc -> DocBook -> XHTML mvn toolchain
In this commit, `mvn -N -Psaxondoc package xml:transform` builds the
AsciiDoc source docs/index.txt into a DocBook file,
`target/generated-docs/index.xml`, which is then transformed with
Saxon using the xslthl extension for syntax highlighting to produce
the XHTML document `target/generated-resources/xml/xslt/index.xml`.

common.xsl is taken from AsciiDoc; htcust.xml is a half-page so-called
"customization layer" stylesheet that glues together common.xsl with
xslthl highlighting.

This is the earliest proof-of-concept stage.  Much work remains in
both tweaking the XHTML style to fit with the new Titan website and in
porting the existing wikidocs to AsciiDoc before this could be
considered ready for distribution.