Switch to HTML 4.01 Transitional and DocBook 5
The current latest version of the AsciiDoctor maven plugin defaults to
DocBook version 4.5.  According to this issue, AsciiDoctor's default
DocBook version is changing to 5.0 in AsciiDoctor 1.5.0 (which has not
yet been released):

  https://github.com/asciidoctor/asciidoctor/issues/554

This commit forces docbook5 generation on the current AsciiDoctor
plugin by setting the backend to "docbook5".  This commit also
modifies htcust.xsl to generate HTML with the 4.01 Transitional
DOCTYPE and set the output encoding to utf-8.  The generated HTML
currently validates on http://validator.w3.org/check.  I also turned
on indentation in the HTML generator.