Add nav footer back into chunked html docs
The nav footer returns -- it's functional again, but the style is
still vanilla and pretty ugly.  The nav header is still out.  Making
it play nice with the sticky/static Titan:db header is going to take
some additional style work that I don't want to get bogged down in
yet.